refactor: separate e2e types from helper functions
Move createE2EEvent to e2e-helpers.ts, keeping e2e-types.ts as pure
type definitions.
